First Launch Page

+++ lib/widgets/product_details_view2.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'dart:io';
 import 'dart:math';
 
 import 'package:adaptive_dialog/adaptive_dialog.dart';
@@ -498,6 +499,7 @@ class _ProductDetailsViewState2 extends State<ProductDetailsView2>  with
   @override
   Widget build(BuildContext contextOver) {
     return Scaffold(
+      resizeToAvoidBottomInset: false,
       backgroundColor: Colors.white,
       body: SafeArea(
         top: true,
@@ -2089,28 +2091,53 @@ class _ProductDetailsViewState2 extends State<ProductDetailsView2>  with
                                                               title: 'Are you sure you want to remove this product?',
                                                               message: 'This action cannot go back later.',
                                                               defaultType: OkCancelAlertDefaultType.cancel,
-                                                            ).then((result) {
+                                                            ).then((result) async {
                                                               if(result == OkCancelResult.ok) {
-                                                                // print(widget.imgUrl);
-                                                                removeImgServer(widget.imgUrl).then((value) async {
-                                                                  print('val lue' + value.toString());
-                                                                  if(value == 200) {
-                                                                    print('all set');
-                                                                    smartKyatFlash(prodName.toString() + ' is successfully removed.', 's');
+                                                                print('widgeturl' + widget.imgUrl);
+                                                                if(widget.imgUrl != null && widget.imgUrl != '') {
+                                                                  try {
+                                                                    final resultInt = await InternetAddress.lookup('google.com');
+                                                                    if (resultInt.isNotEmpty && resultInt[0].rawAddress.isNotEmpty) {
+                                                                      removeImgServer(widget.imgUrl).then((value) async {
+                                                                        print('val lue' + value.toString());
+                                                                        if(value == 200) {
+                                                                          print('all set');
+                                                                          smartKyatFlash(prodName.toString() + ' is successfully removed.', 's');
 
-                                                                    await product.update({
-                                                                      'prods.' + widget.idString: FieldValue.delete()
-                                                                    }).then((value) async {
-                                                                      await productImg.update({
-                                                                        'prods.' + widget.idString: FieldValue.delete()
-                                                                      }).then((value) {
+                                                                          product.update({
+                                                                            'prods.' + widget.idString: FieldValue.delete()
+                                                                          }).then((value) async {
+                                                                          }).catchError((error) => print("Failed to update:  0 $error"));
 
-                                                                      }).catchError((error) => print("Failed to update: 1 $error"));
-                                                                    }).catchError((error) => print("Failed to update:  0 $error"));
+                                                                          productImg.update({
+                                                                            'prods.' + widget.idString: FieldValue.delete()
+                                                                          }).then((value) {
+                                                                          }).catchError((error) => print("Failed to update: 1 $error"));
 
-                                                                    Navigator.pop(contextOver);
+                                                                          Navigator.pop(contextOver);
+                                                                        }
+                                                                      });
+                                                                    }
+                                                                  } on SocketException catch (_) {
+                                                                    smartKyatFMod(context,'Internet connection is required if this product has image.', 'w');
                                                                   }
-                                                                });
+                                                                } else {
+                                                                  print('all set no img');
+                                                                  smartKyatFlash(prodName.toString() + ' is successfully removed.', 's');
+
+                                                                  product.update({
+                                                                    'prods.' + widget.idString: FieldValue.delete()
+                                                                  }).then((value) async {
+                                                                  }).catchError((error) => print("Failed to update:  0 $error"));
+
+                                                                  productImg.update({
+                                                                    'prods.' + widget.idString: FieldValue.delete()
+                                                                  }).then((value) {
+                                                                  }).catchError((error) => print("Failed to update: 1 $error"));
+
+                                                                  Navigator.pop(contextOver);
+                                                                }
+
 
 
                                                               }
@@ -2274,6 +2301,175 @@ class _ProductDetailsViewState2 extends State<ProductDetailsView2>  with
           ),
         ]
     );
+  }
+
+  FlashController? _previousController;
+
+  Future smartKyatFMod<T>(BuildContext context, String message, String type) async {
+    if(_previousController != null) {
+      if (_previousController!.isDisposed == false) _previousController!.dismiss();
+    }
+
+    Widget widgetCon = Container();
+    Color bdColor = Color(0xffffffff);
+    Color bgColor = Color(0xffffffff);
+    if(type == 's') {
+      bdColor = Color(0xffB1D3B1);
+      bgColor = Color(0xffCFEEE0);
+      widgetCon = Container(
+        width: 18,
+        height: 18,
+        decoration: BoxDecoration(
+            borderRadius: BorderRadius.all(
+              Radius.circular(35.0),
+            ),
+            color: Color(0xff419373)),
+        child: Padding(
+          padding: const EdgeInsets.only(right: 1.0),
+          child: Icon(
+            Icons.check_rounded,
+            size: 15,
+            color: Colors.white,
+          ),
+        ),
+      );
+    } else if(type == 'w') {
+      bdColor = Color(0xffF2E0BC);
+      bgColor = Color(0xffFCF4E2);
+      widgetCon = Container(
+        width: 18,
+        height: 18,
+        decoration: BoxDecoration(
+            borderRadius: BorderRadius.all(
+              Radius.circular(35.0),
+            ),
+            color: Color(0xffF5C04A)),
+        child: Padding(
+          padding: const EdgeInsets.only(left: 6.0, top: 1.0),
+          child: Text('!', textScaleFactor: 1, style: TextStyle(fontWeight: FontWeight.w800, color: Colors.white)),
+          // child: Icon(
+          //   Icons.warning_rounded,
+          //   size: 15,
+          //   color: Colors.white,
+          // ),
+        ),
+      );
+    } else if(type == 'e') {
+      bdColor = Color(0xffEAD2C8);
+      bgColor = Color(0xffFAEEEC);
+      widgetCon = Container(
+        width: 18,
+        height: 18,
+        decoration: BoxDecoration(
+            borderRadius: BorderRadius.all(
+              Radius.circular(35.0),
+            ),
+            color: Color(0xffE9625E)),
+        child: Padding(
+          padding: const EdgeInsets.only(left: 0),
+          child: Icon(
+            Icons.close_rounded,
+            size: 15,
+            color: Colors.white,
+          ),
+        ),
+      );
+    } else if(type == 'i') {
+      bdColor = Color(0xffBCCEEA);
+      bgColor = Color(0xffE8EEF9);
+      widgetCon = Container(
+        width: 18,
+        height: 18,
+        decoration: BoxDecoration(
+            borderRadius: BorderRadius.all(
+              Radius.circular(35.0),
+            ),
+            color: Color(0xff4788E2)),
+        child: Padding(
+          padding: const EdgeInsets.only(left: 6.5, top: 1.5),
+          child: Text('i', textScaleFactor: 1, style: TextStyle(fontSize: 13, fontWeight: FontWeight.w800, color: Colors.white,)),
+          // child: Icon(
+          //   Icons.warning_rounded,
+          //   size: 15,
+          //   color: Colors.white,
+          // ),
+        ),
+      );
+    }
+
+    _previousController = FlashController<T>(
+      context,
+      builder: (context, controller) {
+        return Flash(
+          controller: controller,
+          backgroundColor: Colors.transparent,
+          brightness: Brightness.light,
+          // boxShadows: [BoxShadow(blurRadius: 4)],
+          // barrierBlur: 3.0,
+          // barrierColor: Colors.black38,
+          barrierDismissible: true,
+          behavior: FlashBehavior.floating,
+          position: FlashPosition.top,
+          child: Padding(
+            padding: const EdgeInsets.only(
+                top: 93.0, left: 15, right: 15),
+            child: Container(
+              decoration: BoxDecoration(
+                borderRadius: BorderRadius.all(
+                  Radius.circular(10.0),
+                ),
+                color: bgColor,
+                border: Border.all(
+                    color: bdColor,
+                    width: 1.0
+                ),
+              ),
+              child: ListTile(
+                leading: Padding(
+                  padding: const EdgeInsets.only(top: 2.0),
+                  child: widgetCon,
+                ),
+                minLeadingWidth: 15,
+                horizontalTitleGap: 10,
+                minVerticalPadding: 0,
+                title: Padding(
+                  padding: const EdgeInsets.only(top: 15, bottom: 16.3),
+                  child: Container(
+                    child: Text(message, textScaleFactor: 1, overflow: TextOverflow.visible, style: TextStyle(
+                        fontWeight: FontWeight.w400, fontSize: 15, height: 1.2)),
+                  ),
+                ),
+                // subtitle: Text('shit2'),
+                // trailing: Text('GGG',
+                //   style: TextStyle(
+                //     fontSize: 16,
+                //     fontWeight: FontWeight.w500,
+                //   ),),
+              ),
+            ),
+          ),
+        );
+        // return Flash.dialog(
+        //   controller: controller,
+        //   alignment: const Alignment(0, 0.5),
+        //   margin: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
+        //   borderRadius: const BorderRadius.all(Radius.circular(8.0)),
+        //   backgroundColor: Colors.black87,
+        //   child: DefaultTextStyle(
+        //     style: const TextStyle(fontSize: 16.0, color: Colors.white),
+        //     child: Padding(
+        //       padding:
+        //       const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
+        //       child: Text(message),
+        //     ),
+        //   ),
+        // );
+      },
+      duration: const Duration(milliseconds: 2500),
+      persistent: true,
+      transitionDuration: Duration(milliseconds: 300),
+    );
+    return _previousController!.show();
   }
 // changeUnitName2Stock(String split) {
 //   if(split == 'main') {
